+++ src/passken.d.ts
@@ -110,7 +110,7 @@ declare function randomPwd(opts?: Partial<Options>): string;
 declare function randomSecret(length?: number): string;
 declare function sign(iss: number | string, duration: number, type: Type, b64Keys: string[]): string;
 declare function verify(token: string, b64Keys: string[], ignoreExpiration?: boolean): Payload;
-declare function parseBearerToken(authorization: string | undefined): string;
+declare function parseBearer(authorization: string | undefined): string;
 
 export { 
   getSaltRounds,
@@ -126,5 +126,5 @@ export {
   randomSecret,
   sign,
   verify,
-  parseBearerToken,
+  parseBearer,
 };